Ruleset tests: update for WPCS 3.0
The ruleset tests were failing mainly due to https://github.com/WordPress/WordPress-Coding-Standards/pull/2108 which renamed some violation codes that we were excluding.

The other change to account for was the replacement of a WPCS StrictComparisons sniff with a PHPCSExtras Universal sniff. The sniff in PHPCSExtra contains a fixer. As this is a risky fixer, this fixer is turned off for WPCS. The sniff in PHPCSExtra will provide metrics about loose versus strict comparisons.