Files/IncludingFile: allow for more path-returning functions
As per the thread in 740, these functions all return the full path to a file, so these should all be fine.
As things were, use of these in an `include`/`require` statement would throw an unjustified "Absolute include path must be used" warning.

Includes a few extra tests, though I didn't think it would be necessary to have tests with each of these functions.

Fixes 740