Hard-deprecate all JS specific sniffs
Removing these sniffs needs to wait for a new major, as it would be a breaking change, however, in the mean time, we can:
* Exclude the sniffs from both rulesets. This will prevent PHP_CodeSniffer from throwing deprecation notices about these sniffs.
* Deprecate the sniffs themselves, in case users of the VIP-Coding-Standards would be explicitly including any of these sniff from within their own ruleset.

Fixes 836
Related to 442
Related to 552

Issue 442 should be kept as a reminder to remove these sniffs in VIPCS 4.0.
Issues 267, 376, 393, 411, 524, 525, 526, 527, 528, 650, 686, which are all related to these sniffs, should probably be closed as superseded by the deprecation and the sniffs no longer being used by the VIPCS native rulesets.