Fix dangerous use of contents of constants
The value of tokens constants is not guaranteed, so should never be compared against directly.

Either use the token 'code' and compare against the constant (preferred).
Or use the token 'type' and compare against the constant name.
But **_never_** compare against the value of the constant.

Comparing against the 'type' should generally only be needed for tokens which may not exist in older PHPCS/PHP versions.