Docs: remove `@since` tags which are related to WPCS
Some sniffs were originally hosted in the WordPressCS project and later moved to VIPCS. In that move, the WPCS `@since` tags were not scrubbed/updated. This is misleading as the `@since` tags now refer to versions which never existed for VIPCS.

Note: `@since` tags which do refer to VIPCS versions have been left untouched.