Variables/ServerVariables: bug fix - faulty array key determination
The code to find the array index was flawed and could walk beyond the brackets of this array access.

Additionally, array access keys comprised of multiple tokens were not handled correctly.

Includes tests.

Note: WordPressCS has helper functions to retrieve the array access name, but those are marked as internal, which is the reason to introduce a custom function.