WAIT - VIPCS 4.0 | WIP | Various sniffs: fix property visibility
`public` properties in sniffs can be changed from within a PHPCS XML ruleset file.
A number of properties in VIPCS sniffs should never have been `public` to begin with and for most, the format of the property value (multi-level array) cannot be set from within a custom ruleset anyway.

This commit changes the visibility for these properties from `public` to `private`.

Related to issue 234

Note: the `TaxonomyMetaInOptionsSniff::$option_functions` property is not included in this commit as it is deprecated and will be removed.