Hooks/RestrictedHooks: bug fix - false negatives due to incorrect token walking
The `$parameter['start']` key points to the very first token in the parameter, irrespective of what that token is (whitespace, comment, functional).
The `$parameter['end']` key points to the very last token in the parameter, irrespective of what that token is (whitespace, comment, functional).

This means that when walking the tokens of a parameter both the 'start' token, as well as the 'end' token need to be examined.

The code in the `normalize_hook_name_from_parameter()` method handling concatenated hook names did not do this correctly, leading to false negatives.

Fixed now.

Includes test + updating some pre-existing tests which clearly had a trailing space in the parameter to get the test to pass despite this bug.