Classes/DeclarationCompatibility: bug fix - incorrect method name
The sniff would try to find a `register_one()` method declaration and check that, while the method in the parent class is called `_register_one()`.

This would lead to both false positives (if a `register_one()` method would be declared in the child class) and false negatives (if an incompatible `_register_one()` method would be declared in the child class).

Fixed now.

Includes updating the tests.