Classes/DeclarationCompatibility: bug fix - false positives for valid extra parameters
It's perfectly valid for a child class to overload a parent method and to add additional optional parameters to the method.

The sniff did include code to handle this, but the logic of that code block was incorrect.

The old condition basically checked if the 'default' array index existed and the value of the default parameter was not `true`. Not sure where that came from, but it makes no sense.

Fixed now.

Includes adding support for additional optional parameters being declared as variadic (PHP 5.6+), which makes them optional by nature.
Includes minor efficiency fix - no need to continue examining extra parameters to see if they are optional if we already saw one which was not.

Includes tests.